requirements.txt
Django==6.0.3
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pandas==3.0.1
watchfiles==0.19.0
numpy==2.4.2
pandas==2.3.2
scipy==1.16.2
scikit-learn==1.8.0
statsmodels==0.14.5
imbalanced-learn==0.14.0
xgboost==3.0.5
lightgbm==4.6.0
matplotlib==3.10.8
optuna==4.5.0
mlflow==3.3.2
pyarrow==23.0.1